python: separate bindings for viewing job records
Problem: The code for both viewing job records in the flux-accounting
database and calculating/updating job usage values is in the same file,
job_archive_interface.py. This causes the file to be very large and hard
to read; it would be best to separate these two different features
(viewing job records & calculating usage) into two, smaller files.

Move the code for viewing job records into a new file called
jobs_table_subcommands.py.

Adjust the code in job_archive_interface.py, the flux-accounting
service, and t1006_job_archive.py to account for the newly-moved code.
