+++ src/cmd/flux-account-service.py
@@ -24,6 +24,7 @@ from fluxacct.accounting import bank_subcommands as b
 from fluxacct.accounting import job_archive_interface as jobs
 from fluxacct.accounting import queue_subcommands as qu
 from fluxacct.accounting import project_subcommands as p
+from fluxacct.accounting import jobs_table_subcommands as j
 
 
 def establish_sqlite_connection(path):
@@ -322,7 +323,7 @@ class AccountingService:
     # pylint: disable=no-self-use
     def view_job_records(self, handle, watcher, msg, arg):
         try:
-            val = jobs.output_job_records(
+            val = j.output_job_records(
                 self.conn,
                 msg.payload["output_file"],
                 jobid=msg.payload["jobid"],
